+++ src/BenGorUser/User/Application/Command/PurgeOutdatedTokens/PurgeOutdatedRememberPasswordTokensUserCommand.php
@@ -1,0 +1,22 @@
+<?php
+
+/*
+ * This file is part of the BenGorUser package.
+ *
+ * (c) Beñat Espiña <benatespina@gmail.com>
+ * (c) Gorka Laucirica <gorka.lauzirika@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace BenGorUser\User\Application\Command\PurgeOutdatedTokens;
+
+/**
+ * Purge outdated remember password tokens command class.
+ *
+ * @author Beñat Espiña <benatespina@gmail.com>
+ */
+class PurgeOutdatedRememberPasswordTokensUserCommand
+{
+}